requirements.txt
# Core Dependencies
PyGithub==2.1.1
pandas==2.1.4
python-dotenv==1.0.0
requests==2.31.0

# Development Dependencies
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.7.1

# Data Processing
numpy==1.26.2
pytz==2023.3
